ML/mlintro/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gdown>=4.0.0
librosa>=0.8.1
opencv-python-headless>=4.5.4
mediapipe>=0.8.9
whisper>=1.0.0
deepface>=0.0.79
nltk>=3.6.7
language-tool-python>=2.6.0
fpdf2>=2.5.6
python-multipart>=0.0.5
slowapi>=0.1.5
python-dotenv>=0.19.0
psutil>=5.9.0
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
pydantic>=1.9.0